+++ CHANGELOG.md
@@ -7,8 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `trace` log of enqueued tasks. [himalaya#518]
+
 ## [0.2.2] - 2024-12-09
 
 Starting from this release, we maintain the git history and the changelogs as clean as possible. The documentation will come in the nearest releases.
 
 [0.2.2]: https://github.com/pimalaya/imap-client/releases/tag/v0.2.2
+
+[himalaya#518]: https://github.com/pimalaya/himalaya/issues/518